Add XSLT that blocks messages with source AG123.
Added script that can be used with daffodil CLI
parse command with the stream option to parse
a stream of messages, and transform them all
with XSLT.

    daffodil parse --stream      \
      -s ../src/fakeTDL.dfdl.xsd \
      test_file_01.dat |         \
      ./xslt-stream.sh block_fakeTDL_ag123.xsl

Updated README.md to explain testing with the XSLT transformation.

Removed fakeTDLFile element and all related tests and files.
The use of daffodil CLI with --stream option replaces this for parsing
files of back-to-back messages.